work towards multithreaded and concurrent UI view update loops
tricky part is to get the interaction to work within the update loop. In order for it to be responsive interactions need to block the renderer from requesting new frames and then cause an update right after the interaction is done to not cause glitches (especially when placing in items into the inventory).